vm: baby steps towards the full implementation
this commit adds the alloc::Allocator im using to make sure the compiled byte code uses a free register and keeps track of used registers.
The commit also adds the alloc::Pool for 'allocating' constants and passing them to the vm. A first compilation is done for the expr::Number structure with the expr::Node trait.
I also added the object / type system for the language in types as well as the operations in vm::Operation, this module also includes the implementation of the registerbased virtual machine for the generated byte code. In the main one can see the addition of Operation::Debug and the argument of 0, for debugging/printing the value stored in the register 0 of the virtual machine.